requirements.txt
pypdf
langchain-google-genailanggraph 
langchain-community
tavily-python 
pandas 
tiktoken 
langchain
chromadb
langchain-text-splitters
python-dotenv
langchain-cohere
sentence-transformers
lxml
beautifulsoup4
elevenlabs
assemblyai[extras]
streamlit_webrtc
pyaudio